fix: remove speakeasy docs generate command
